+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [\#66](https://github.com/osiegmar/logback-gelf/issues/66)
 - Replace blank hostname and blank log message to prevent Graylog error
   [\#82](https://github.com/osiegmar/logback-gelf/issues/82)
+- Default of `includeMarker` changed to `false`. Serialization format of markers has changed.
 
 ## [4.0.2] - 2021-12-22
 ### Changed